fixing public folder bug

+++ utils/publicFolder.js
@@ -9,13 +9,15 @@ const publicFolder = asyncHandeler(async (req, res, next) => {
 
   fs.readdir(path.join(__dirname, "..", "public"), (err, files) => {
     if (err) console.log(err);
-
-    for (const file of files) {
-      fs.unlink(path.join(__dirname, "..", "public", file), (err) => {
-        if (err) console.log(err);
-      });
+    if (files.length > 0) {
+      for (const file of files) {
+        fs.unlink(path.join(__dirname, "..", "public", file), (err) => {
+          if (err) console.log(err);
+        });
+      }
     }
-
+    fs.rmdirSync(path.join(__dirname, "..", "public"));
+    fs.mkdirSync(path.join(__dirname, "..", "public"));
     res
       .status(200)
       .json({ success: true, message: "folder refreshed sucesfully" });
